pyproject.toml
[tool.poetry]
name = "sugatan_hyros"
version = "0.1.0"
description = ""
authors = ["HM <hieumdd@gmail.com>"]

[tool.poetry.dependencies]
python = "~3.9"
aiohttp = "^3.7.4"
Jinja2 = "^3.0.1"
google-cloud-bigquery = "^2.28.0"
google-cloud-pubsub = "^2.8.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
black = "^21.9b0"
pytest-timeout = "^1.4.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "-s --durations=0"
timeout = 530
testpaths = [
    "test",
]
